Struct kodak::Image

pub struct Image { /* private fields */ }

The Image struct is at the heart of Kodak. You’ll be using functions on this 95% of the time.

Note that the maximum image size is 2^32 - 1 by 2^32 - 1 pixels. This limit was chosen because it is also the maximum of the PNG format.

Implementations

impl Image

pub fn blank(dimension: Dim) -> Self

Creates a new blank image.

The default colour used is black.

Arguments

dimension - the dimensions of the image to be created

Returns

The newly created image.

impl Image

pub fn get_dimensions(&self) -> Dim

Returns the dimensions of the image.

pub fn as_region(&self) -> Region

Returns the entire image as a region.

pub fn get_pixel(&self, loc: Loc) -> Result<Colour, &'static str>

Tries to look up the colour of a specific pixel; returns an Err<&str> if the location is out of bounds and an Ok if not.

impl Image

pub fn fill(self, colour: Colour) -> Image

Fill the entire image with a given colour.

Arguments

colour: the colour to fill with.

Examples

let img = Image::blank(30, 20).fill(Colour::WHITE);
assert_eq!(img.pixels[0], Colour::WHITE);

pub fn fill_region(self, region: Region, colour: Colour) -> Image

Fills a region.

pub fn crop_unclamped(self, region: Region) -> Image

Crop a region out of the image and return it. This method is mainly used internally and panics; the safer crop() should be used instead.

The corner from which to crop is assumed to be the top left corner.

Panics

if the corner from which to crop is outside of the image,

if the cropped image would reach outside of the image (by specifying new dimensions which are too large).

pub fn crop(self, region: Region) -> Result<Image, &'static str>

Crop a region out of the image and return it. This method (unlike crop_unclamped()) will adjust the size of the region if it is too big. It will return an Err<&str> if the corner to begin with falls outside of the image.

Arguments

region - the region to crop out (inclusive).

Examples

let img = Image::blank( Dim { w: 20, h: 30 } )
    .crop( Region { l: Loc { x: 10, y: 10 }, d: Dim { w: 10, h: 10 } } )
    .unwrap();
assert_eq!(img.width, 10);